base_components/data_verify_storge/requirements.txt
elasticsearch==8.19.2
python-dotenv
pika>=1.3.0